requirements.txt
pygame
taichi
PyOpenGL
PyOpenGL_accelerate